By default only load session used the last year
Older students can restart their work, but they will not
be included in the statistics or exported CSV grades until
they reconnect to the session.

The Grader role may trigger the loading of all the students.

The inactive student time may be modified on the option page.
The may trigger the loading of new students (no one is removed)