EXPERIMENTAL: Quenlig is now threaded
In order to use the Author role with 500 students it
is necessary to use threads to not froze students when the teacher
load big stat pages.

The function updating data, run locked:

    question_stats(), get_state(), student()

The 'student' object is locked by any request from any of its sessions.

The grade_recompute function lock recomputed students.